dd439dc Jeff Davidson on 10/16/15 at 06:26 (committed on 3/11/16 at 05:49) Retry more errors in Volley's BasicNetwork.
Always retry I/O errors while reading the HTTP response entity.
Furthermore, if a client opts in, retry HTTP 500 errors indicating
something went wrong on the server.

This resolves a longstanding TODO to only throw a ServerError on 5xx
errors by adding a ClientError for 4xx errors. For backwards
compatibility, this extends ServerError.

Note that Volley already may retry a request that reached the server
if the connection times out, which means that lack of idempotency
shouldn't be a concern here if it wasn't already. But if we wanted to
be even safer, we could require clients to opt into the additional
retry cases, at the cost of a somewhat more polluted API.

Bug: 23152983
Change-Id: I92cf35c66ccf98a1682adf41654afeb8634911db

(cherry picked from commit dd439dcb42093fde4452e320f3a9c4bfa2066def of volley)